requirements.txt

streamlit
yfinance
newspaper3k
vaderSentiment
feedparser
beautifulsoup4
python-dotenv
matplotlib
